Working on improving gauge sizing/resizing in create
With a big window its fine, but as soon as moving to a small window, everything gets messed up.
Attempting to enforce widths based on the relative 'size' of the gauge. Gets tricky as the
window resizes and we have to still modify the size of the gauge accordingly to fit well.

This is probably going to be a fairly invasive fix, requiring digging into css compiled for
gauges across all the dashboard instances. Also needs testing with multiple gauges and
window sizes